Moved a few things from web installer to base installer
