+++ mountequist/installers/web.py
@@ -1,25 +1,17 @@
 import os
-import zipfile
 
 import requests
 
 from mountequist.exceptions import InstallError
 from mountequist.installers.base import Installer
-from mountequist.util import get_root_mountebank_path
 
 WINDOWS_LINK = (
     "https://s3.amazonaws.com/mountebank/v1.14/mountebank-v1.14.0-win-x64.zip")
-
-NODE_FILENAME = "node.exe"
 
 
 class WindowsWeb(Installer):
     def __init__(self, link=None):
         self.link = link if link is not None else WINDOWS_LINK
-
-    @classmethod
-    def is_installed(cls, mountebank_path):
-        return cls.find_exe(mountebank_path) is not None
 
     def install(self, mountebank_path):
         if not os.path.exists(mountebank_path):
@@ -47,23 +39,8 @@ class WindowsWeb(Installer):
         return zip_file_path
 
     @classmethod
-    def _extract(cls, zip_file_path, mountebank_path):
-        with zipfile.ZipFile(zip_file_path) as zip_file:
-            zip_file.extractall(mountebank_path)
-
-    @classmethod
     def _install(cls, link, mountebank_path):
         zip_file_path = cls._download(link, mountebank_path)
         cls._extract(zip_file_path, mountebank_path)
 
         return cls.find_exe(mountebank_path)
-
-    @classmethod
-    def find_exe(cls, mountebank_path):
-        elements = os.listdir(mountebank_path)
-        if not elements:
-            return
-
-        root_path = get_root_mountebank_path(mountebank_path)
-        if NODE_FILENAME in os.listdir(root_path):
-            return os.path.join(root_path, NODE_FILENAME)
